prepare for module use
