+++ go.mod
@@ -4,5 +4,6 @@ require (
 	github.com/armon/go-socks5 v0.0.0-20160902184237-e75332964ef5
 	github.com/gobuffalo/packr v1.21.9
 	github.com/gobuffalo/packr/v2 v2.0.0-rc.15
+	github.com/spf13/afero v1.2.0 // indirect
 	golang.org/x/crypto v0.0.0-20190123085648-057139ce5d2b
 )
